Make sure to cast time_t to i64 so that notmuch-rs works on 32 bit platforms too

+++ src/message.rs
@@ -86,7 +86,7 @@ where
     }
 
     pub fn date(&self) -> i64 {
-        unsafe { ffi::notmuch_message_get_date(self.handle.ptr) }
+        unsafe { ffi::notmuch_message_get_date(self.handle.ptr) as i64 }
     }
 
     pub fn header(&self, name: &str) -> Result<Option<&str>> {
